moving edge case to own test module
